+++ src/web/Models/FileHandleConfig.cs
@@ -1,4 +1,7 @@
 namespace web.Models;
 
-public record FileHandleConfig(string path, FileMode fileMode, FileAccess FileAccess,
-        FileShare fileShare, uint preallocationSize);
+public record FileHandleConfig(string path,
+                               FileMode fileMode,
+                               FileAccess FileAccess,
+                               FileShare fileShare,
+                               long preallocationSize);
